Studia/BGT/lab1/requirements.txt
Flask==1.1.2
google-cloud-firestore==1.6.2
google-cloud-tasks==1.5.0
